I added in the full implementation of the TB661Motor class. However, I can't get it working. The process freezes when I try to access the pins.

+++ MotorLibrary/MotorLibrary/src/tweeter0830/motorLibrary/DcMotor.java
@@ -1,9 +1,9 @@
 package tweeter0830.motorLibrary;
 
 public interface DcMotor {
-	void moveForward(float speed);
-	void moveBackward(float speed);
-	void move(float speed);
+	void moveForward(double speed);
+	void moveBackward(double speed);
+	void move(double speed);
 	void brake();
 	void standby();
 	void powerOn();
